+++ lib/punchblock/component/input.rb
@@ -36,9 +36,15 @@ module Punchblock
 
       # @return [Integer] Indicates (in the case of DTMF input) the amount of time (in milliseconds) between input digits which may expire before a timeout is triggered.
       attribute :inter_digit_timeout, Integer
+      #
+      # @return [Integer] Indicates (in the case of DTMF input) the amount of time between input digits which may expire before a match is returned.
+      attribute :dtmf_term_timeout, Integer
 
       # @return [Integer] Indicates the amount of time during input that recognition will occur before a timeout is triggered.
       attribute :recognition_timeout, Integer
+
+      # @return [Integer] Indicates the length of time additional input may be accepted following a match before returning the result.
+      attribute :speech_complete_timeout, Integer 
 
       attribute :grammars, Array, default: []
       def grammars=(others)
@@ -72,7 +78,9 @@ module Punchblock
           'sensitivity' => sensitivity,
           'initial-timeout' => initial_timeout,
           'inter-digit-timeout' => inter_digit_timeout,
-          'recognition-timeout' => recognition_timeout
+          'dtmf-term-timeout' => dtmf_term_timeout,
+          'recognition-timeout' => recognition_timeout,
+          'speech-complete-timeout' => speech_complete_timeout
         }
       end