+++ spec/punchblock/translator/asterisk/component/mrcp_prompt_spec.rb
@@ -133,20 +133,33 @@ module Punchblock
             context 'with multiple inline documents' do
               let(:output_command_options) { { render_documents: [{value: ssml_doc}, {value: ssml_doc}] } }
 
-              it "should return an error and not execute any actions" do
+              it "should return a ref and execute SynthAndRecog" do
+                param = [[ssml_doc, ssml_doc].map{|d| d.to_doc.to_s.squish}.join('^'), voice_grammar.to_doc.to_s].map { |o| "\"#{o.to_s.squish.gsub('"', '\"')}\"" }.push('uer=1&b=1').join(',')
+                mock_call.should_receive(:execute_agi_command).once.with('EXEC SynthAndRecog', param).and_return code: 200, result: 1
                 subject.execute
-                error = ProtocolError.new.setup 'option error', 'Only one document is allowed.'
-                expect(original_command.response(0.1)).to eq(error)
+                original_command.response(0.1).should be_a Ref
               end
             end
 
             context 'with multiple documents by URI' do
               let(:output_command_options) { { render_documents: [{url: 'http://example.com/doc1.ssml'}, {url: 'http://example.com/doc2.ssml'}] } }
 
-              it "should return an error and not execute any actions" do
+              it "should return a ref and execute SynthAndRecog" do
+                param = ['http://example.com/doc1.ssml^http://example.com/doc2.ssml', voice_grammar.to_doc.to_s].map { |o| "\"#{o.to_s.squish.gsub('"', '\"')}\"" }.push('uer=1&b=1').join(',')
+                mock_call.should_receive(:execute_agi_command).once.with('EXEC SynthAndRecog', param).and_return code: 200, result: 1
                 subject.execute
-                error = ProtocolError.new.setup 'option error', 'Only one document is allowed.'
-                expect(original_command.response(0.1)).to eq(error)
+                original_command.response(0.1).should be_a Ref
+              end
+            end
+
+            context 'with audiofile document' do
+              let(:output_command_options) { { render_documents: [{url: '/filesystem/upload.wav', content_type: 'audio/wav'}, {url: '/filesystem/recording.ulaw', content_type: 'audio/ulaw'}] } }
+
+              it "should return a ref and execute SynthAndRecog" do
+                param = ['audio:/filesystem/upload^audio:/filesystem/recording', voice_grammar.to_doc.to_s].map { |o| "\"#{o.to_s.squish.gsub('"', '\"')}\"" }.push('uer=1&b=1').join(',')
+                mock_call.should_receive(:execute_agi_command).once.with('EXEC SynthAndRecog', param).and_return code: 200, result: 1
+                subject.execute
+                original_command.response(0.1).should be_a Ref
               end
             end
 
@@ -162,7 +175,7 @@ module Punchblock
           end
 
           describe 'Output#renderer' do
-            [nil, :unimrcp].each do |renderer|
+            [nil, :unimrcp, :native_or_unimrcp].each do |renderer|
               context renderer.to_s do
                 let(:output_command_opts) { { renderer: renderer } }
 
@@ -682,6 +695,24 @@ module Punchblock
 
             context 'unset' do
               let(:input_command_opts) { { headers: {"Speech-Complete-Timeout" => nil } } }
+
+              it 'should not pass any options to SynthAndRecog' do
+                expect_synthandrecog_with_options(//)
+                subject.execute
+              end
+            end
+
+            context 'a positive number (non-header)' do
+              let(:input_command_opts) { { speech_complete_timeout: 1000 } }
+
+              it 'should pass the dit option to SynthAndRecog' do
+                expect_synthandrecog_with_options(/sct=1000/)
+                subject.execute
+              end
+            end
+
+            context 'unset (non-header)' do
+              let(:input_command_opts) { { speech_complete_timeout: nil } }
 
               it 'should not pass any options to SynthAndRecog' do
                 expect_synthandrecog_with_options(//)