+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [ifbyphone-develop](https://github.com/cloudvox/punchblock)
+  * Feature: Support `:native_or_unimrcp` renderer when using `:unimrcp` recognizer for ASR
+  * Bugfix: Restore mainline defaults for UniMRCP settings
+  * Feature: Allow passing multiple ssml documents and audiofiles to UniMRCP (requires UniMRCP revision >= r2153)
+  * Feature: Enable setting 'Dtmf-Term-Timeout' value for UniMRCP; necessary to exit sooner from successful Lumenvox recognition event
+  * Feature: Enable setting 'Speech-Complete-Timeout' value for UniMRCP; necessary to allow additional input after an initial match has been recognized
+
 # [develop](https://github.com/adhearsion/punchblock)
 
 # [v2.7.5](https://github.com/adhearsion/punchblock/compare/v2.7.4...v2.7.5) - [2015-08-27](https://rubygems.org/gems/punchblock/versions/2.7.5)